refactor(daemon): route process discovery through ports.ProcessObserver (#478)
Concentrate the OS coupling of process discovery behind a single seam so
adding a platform is "implement one interface in one file" rather than
touching every discovery path.

  - Add outbound.ProcessObserver (FindByName/FindByCmdline/CWDOf/WriterOf/
    EnvOf) next to ProcessWatcher.
  - process_darwin.go — the macOS pgrep/lsof/sysctl mechanism, moved verbatim
    behind darwinObserver. newObserver() picks it at compile time.
  - process_nondarwin.go — stub observer so the package builds on every other
    OS; env still works via the existing build-tagged readProcessEnv. Replaced
    by a real /proc impl in the next phase.
  - discovery.go / scanner.go / osutil.go now call osProc.* instead of the
    package-level pgrep/lsof helpers; those helpers moved into the observer.

Pure refactor — darwin behaviour is byte-for-byte unchanged: full
`go test ./core/... -race` stays green and replay goldens are untouched
(replay never imports processlifecycle). Linux and the windows stub compile.
